requirements.txt
asgiref==3.7.2
crispy-bootstrap4==2023.1
Django==5.0
django-crispy-forms==2.1
django-tables2==2.7.0
gunicorn==21.2.0
mysqlclient==2.2.0
packaging==23.2
Pillow==10.1.0
python-dotenv==1.0.0
sqlparse==0.4.4
whitenoise==6.6.0
django-modeltranslation==0.18.11